Refactor: MainTitleView 페이지 분리
